refactor postcode and heirarchy into property model
